wasm has its own file
