+++ packages/bun-types/index.d.ts
@@ -16,3 +16,4 @@
 /// <reference path="./html-rewriter.d.ts" />
 /// <reference path="./jsc.d.ts" />
 /// <reference path="./sqlite.d.ts" />
+/// <reference path="./wasm.d.ts" />
